made miliseconds a long

+++ include/sound.c
@@ -10,7 +10,7 @@
 
 int con = -1;
 
-void wait(double miliseconds) {
+void wait(long miliseconds) {
     tv.secs = (int) (miliseconds / 1000);
     tv.usecs = (int) (miliseconds % 1000);
     select(0, NULL, NULL, NULL, &tv);
